requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain==1.0.8
langchain-ollama
langchain-community
langchain-core
langchain-text-splitters
chromadb
pypdf
